+++ x/evidence/testutil/expected_keepers_mocks.go
@@ -11,7 +11,6 @@ import (
 
 	stakingv1beta1 "cosmossdk.io/api/cosmos/staking/v1beta1"
 	address "cosmossdk.io/core/address"
-	comet "cosmossdk.io/core/comet"
 	math "cosmossdk.io/math"
 	types "github.com/cosmos/cosmos-sdk/crypto/types"
 	types0 "github.com/cosmos/cosmos-sdk/types"
@@ -54,21 +53,6 @@ func (m *MockStakingKeeper) ConsensusAddressCodec() address.Codec {
 func (mr *MockStakingKeeperMockRecorder) ConsensusAddressCodec() *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ConsensusAddressCodec", reflect.TypeOf((*MockStakingKeeper)(nil).ConsensusAddressCodec))
-}
-
-// GetParams mocks base method.
-func (m *MockStakingKeeper) GetParams(ctx context.Context) (types1.Params, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GetParams", ctx)
-	ret0, _ := ret[0].(types1.Params)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// GetParams indicates an expected call of GetParams.
-func (mr *MockStakingKeeperMockRecorder) GetParams(ctx interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetParams", reflect.TypeOf((*MockStakingKeeper)(nil).GetParams), ctx)
 }
 
 // ValidatorByConsAddr mocks base method.
@@ -335,41 +319,4 @@ func (m *MockBankKeeper) SendCoinsFromModuleToAccount(ctx types0.Context, sender
 func (mr *MockBankKeeperMockRecorder) SendCoinsFromModuleToAccount(ctx, senderModule, recipientAddr, amt interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SendCoinsFromModuleToAccount", reflect.TypeOf((*MockBankKeeper)(nil).SendCoinsFromModuleToAccount), ctx, senderModule, recipientAddr, amt)
-}
-
-// MockCometinfo is a mock of Cometinfo interface.
-type MockCometinfo struct {
-	ctrl     *gomock.Controller
-	recorder *MockCometinfoMockRecorder
-}
-
-// MockCometinfoMockRecorder is the mock recorder for MockCometinfo.
-type MockCometinfoMockRecorder struct {
-	mock *MockCometinfo
-}
-
-// NewMockCometinfo creates a new mock instance.
-func NewMockCometinfo(ctrl *gomock.Controller) *MockCometinfo {
-	mock := &MockCometinfo{ctrl: ctrl}
-	mock.recorder = &MockCometinfoMockRecorder{mock}
-	return mock
-}
-
-// EXPECT returns an object that allows the caller to indicate expected use.
-func (m *MockCometinfo) EXPECT() *MockCometinfoMockRecorder {
-	return m.recorder
-}
-
-// GetCometBlockInfo mocks base method.
-func (m *MockCometinfo) GetCometBlockInfo(arg0 context.Context) comet.BlockInfo {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GetCometBlockInfo", arg0)
-	ret0, _ := ret[0].(comet.BlockInfo)
-	return ret0
-}
-
-// GetCometBlockInfo indicates an expected call of GetCometBlockInfo.
-func (mr *MockCometinfoMockRecorder) GetCometBlockInfo(arg0 interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetCometBlockInfo", reflect.TypeOf((*MockCometinfo)(nil).GetCometBlockInfo), arg0)
 }