+++ x/evidence/types/evidence_test.go
@@ -75,64 +75,21 @@ func TestEquivocationValidateBasic(t *testing.T) {
 func TestEvidenceAddressConversion(t *testing.T) {
 	sdk.GetConfig().SetBech32PrefixForConsensusNode("testcnclcons", "testcnclconspub")
 	tmEvidence := NewCometMisbehavior(1, 100, time.Now(), comet.DuplicateVote,
-		validator{address: []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0}, power: 100})
+		comet.Validator{Address: []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0}, Power: 100})
 
 	evidence := types.FromABCIEvidence(tmEvidence, address.NewBech32Codec("testcnclcons"))
 	consAddr := evidence.GetConsensusAddress(address.NewBech32Codec("testcnclcons"))
 	// Check the address is the same after conversion
-	require.Equal(t, tmEvidence.Validator().Address(), consAddr.Bytes())
+	require.Equal(t, tmEvidence.Validator.Address, consAddr.Bytes())
 	sdk.GetConfig().SetBech32PrefixForConsensusNode(sdk.Bech32PrefixConsAddr, sdk.Bech32PrefixConsPub)
 }
 
-type Misbehavior struct {
-	height           int64
-	time             time.Time
-	totalVotingPower int64
-	validator        validator
-	misBehaviorType  comet.MisbehaviorType
-}
-
-func NewCometMisbehavior(height, tvp int64, t time.Time, tpe comet.MisbehaviorType, val validator) comet.Evidence {
-	return Misbehavior{
-		height:           height,
-		time:             t,
-		totalVotingPower: tvp,
-		misBehaviorType:  tpe,
-		validator:        val,
+func NewCometMisbehavior(height, tvp int64, t time.Time, tpe comet.MisbehaviorType, val comet.Validator) comet.Evidence {
+	return comet.Evidence{
+		Height:           height,
+		Time:             t,
+		TotalVotingPower: tvp,
+		Type:             tpe,
+		Validator:        val,
 	}
-}
-
-func (m Misbehavior) Type() comet.MisbehaviorType {
-	return m.misBehaviorType
-}
-
-func (m Misbehavior) Height() int64 {
-	return m.height
-}
-
-func (m Misbehavior) Validator() comet.Validator {
-	return m.validator
-}
-
-func (m Misbehavior) Time() time.Time {
-	return m.time
-}
-
-func (m Misbehavior) TotalVotingPower() int64 {
-	return m.totalVotingPower
-}
-
-type validator struct {
-	address []byte
-	power   int64
-}
-
-var _ comet.Validator = (*validator)(nil)
-
-func (v validator) Address() []byte {
-	return v.address
-}
-
-func (v validator) Power() int64 {
-	return v.power
 }